:bug: sns: Protect topic from deletion in a simpler, more specific way
Previously, a `aws_sns_topic_policy` resource was created for the topic policy.
When running terraform destroy or deleting the sns topic from terraform
source code, the topic policy would have been deleted first and then, because
no policy would protect it anymore, also the topic.

To prevent this, sns:SetTopicAttributes needed to be also Denied. But this
was brittle and ugly because it prevents also setting other attributes and
requires manual interactions instead of just repeated config adjustments
to code running in the CI pipeline if you really want to delete a protected
topic after all.

Now, the policy will be set as an attribute of the `aws_sns_topic` resource,
thereby making it impossible to accidentally delete first the policy and then
the topic. Since the policy is part of the topic, they would be deleted
at the same time, but this is prevented by the Deny of sns:DeleteTopic.